Refactored to use Sass.
