+++ client/src/components/MainProductDetail/MainProductDetailS.css
@@ -1,0 +1,4 @@
+.MainProductDetail {
+  background-color: whitesmoke;
+  padding: 10px;
+}
