+++ client/src/components/UserLists/UserLists.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import './UserListsS.css';
 
 function listChecker(list, itemNum) {
   for (let i = 0; i < list.length; i += 1) {
