+++ client/src/components/functionalComps/functionalCompsS.css
@@ -1,0 +1,6 @@
+.ExclusiveTag {
+  display: inline-block;
+  background-color: yellow;
+  padding: 5px;
+  margin: 10px;
+}
