+++ client/src/indexS.css
@@ -1,0 +1,8 @@
+body {
+  background-color: grey;
+}
+
+#root{
+  display: flex;
+  justify-content: center;
+}
