--- client/src/styles.css
@@ -1,44 +1,0 @@
-body {
-  background-color: grey;
-}
-
-.MainProductDetail {
-  background-color: whitesmoke;
-  padding: 10px;
-}
-
-.ExclusiveTag {
-  display: inline-block;
-  background-color: yellow;
-  padding: 5px;
-  margin: 10px;
-}
-
-.SeriesImage {
-  display: inline;
-  max-inline-size: 100px;
-}
-
-.RatingStars {
-  display: inline-block;
-}
-
-.ReviewCount {
-  display: inline-block;
-  margin: 10px;
-}
-
-.Star {
-  display: inline-block;
-  max-inline-size: 30px;
-}
-
-.WishListHeart {
-  display: inline-block;
-  max-inline-size: 40px;
-}
-
-.EmailIcon {
-  display: inline-block;
-  max-inline-size: 40px;
-}
